pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "earnings-alpha"
version = "0.1.0"
description = "NLP-derived alpha signals from earnings call transcripts"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Earnings Alpha Contributors" }]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "yfinance>=0.2.33",
    "pandas-datareader>=0.10.0",
    "praw>=7.7.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.0",
    "bertopic>=0.16.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.33",
    "linearmodels>=5.3",
    "statsmodels>=0.14.0",
    "pytorch-forecasting>=1.0.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "shap>=0.43.0",
    "mlflow>=2.9.0",
    "prefect>=2.14.0",
    "vectorbt>=0.26.0",
    "streamlit>=1.29.0",
    "pyarrow>=14.0.0",
    "duckdb>=0.9.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]

[project.scripts]

[tool.setuptools.packages.find]
where = ["."]
include = ["ingestion*", "nlp*", "causal*", "forecasting*", "explainability*", "backtesting*", "mlops*", "dashboard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
